refactor: /utils index.ts 추가
