+++ insty-frontend/src/app/api/interceptor.ts
@@ -1,5 +1,7 @@
 import axios from "axios";
 
+import { getAccessToken } from "@/app/utils";
+
 const axiosInstance = axios.create({
 	baseURL: process.env.NEXT_PUBLIC_BACK_BASE_URL,
 	withCredentials: true,
@@ -8,7 +10,7 @@ const axiosInstance = axios.create({
 axiosInstance.interceptors.request.use(
 	(config) => {
 		if (typeof window !== "undefined") {
-			const accessToken = localStorage.getItem("accessToken");
+			const accessToken = getAccessToken();
 			if (accessToken) {
 				config.headers.Authorization = `Bearer ${accessToken}`;
 			}
