+++ insty-frontend/src/app/utils/index.ts
@@ -1,0 +1,3 @@
+export * from "./date";
+export * from "./regex";
+export * from "./storage";
